requirements.txt
Django==4.2.13
djangorestframework==3.15.2
mongoengine==0.29.1
django-rest-framework-mongoengine==3.4.1
PyJWT==2.9.0
python-dotenv==1.0.1
bcrypt==4.2.0